[DH-491] Move audit serialiser to separate module
